+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- New UI: Metadata Retriever
+
 ## [6.13.0] 2024-10-19
 
 - Live DevOps Pipelines